chore: enforce `TasksMax` to `infinity`
Specify the maximum number of tasks that may be created in the unit.
This ensures that the number of tasks accounted for the unit stays
below a specific limit. This either takes an absolute number of tasks or a
percentage value that is taken relative to the configured maximum number of
tasks on the system. If assigned the special value "infinity", no tasks limit
is applied. This controls the "pids.max" control group attribute.

See also: https://www.nomadproject.io/guides/install/production/deployment-guide.html#configure-systemd

Please note, TasksMax is only available in systemd >= 226