Improve Vault Support for Namespace, Tokens
This adds support for Vault namespaces (from Vault Enterprise).
Additionally, this ensures that a vault token is not installed on client
nodes. Per the nomad documentation, this token only needs be installed
on servers, and tokens will be delegated appropriately to client nodes.

Addresses https://github.com/ansible-community/ansible-nomad/issues/101