Adding support for JWT auth to vault
In newer versions of nomad you get this warning:

```
agent: Setting a Vault token in the agent configuration is deprecated and will be removed in Nomad 1.9. Migrate your Vault configuration to use workload identity.
```

This PR adds a default idenity section the config if set, so that we
don't break older nomad versions.